refactor(react): rewrite stack
