+++ packages/react/src/layout/inline/Inline.stories.tsx
@@ -1,66 +1,38 @@
 import * as React from 'react';
+import { Story } from '@storybook/react';
 
-import { SystemBlock, ComponentBlock } from '../../utils/storybook';
 import { Box } from '../box';
-import { Card } from '../card';
 import { Text } from '../typography';
-import { Stack } from '../stack';
-import { Inline } from '.';
-
-const readme = require('./README.md');
+import Inline, { InlineProps } from './Inline';
 
 export default {
   title: 'Core/Layout/Inline',
   component: Inline,
-  parameters: {
-    notes: { markdown: readme },
-    jsx: { skip: 2 },
+  argTypes: {
+    spacing: {
+      options: ['xxxs', 'xxs', 'xs', 'sm', 'md', 'lg', 'xl', 'xxl'],
+    },
   },
 };
 
-export const Example = () => {
+export const Example: Story<InlineProps> = ({ spacing }) => {
   return (
-    <SystemBlock
-      title="Inline"
-      subtitle="Use the Inline component to evenly space inner elements horizontally"
-      noBackground
-    >
-      <Stack spacing="xl">
-        <ComponentBlock title="Inline (spacing=xs)">
-          <Card padding="sm" elevation={2}>
-            <Box backgroundColor="red02" overflow="hidden">
-              <Inline spacing="xs">
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-              </Inline>
-            </Box>
-          </Card>
-        </ComponentBlock>
-        <ComponentBlock title="Inline (spacing=sm)">
-          <Card padding="sm" elevation={2}>
-            <Box backgroundColor="red02" overflow="hidden">
-              <Inline spacing="sm">
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-              </Inline>
-            </Box>
-          </Card>
-        </ComponentBlock>
-      </Stack>
-    </SystemBlock>
+    <Box backgroundColor="red02">
+      <Inline spacing={spacing}>
+        <Box backgroundColor="white">
+          <Text>Box (backgroundColor=white)</Text>
+        </Box>
+        <Box backgroundColor="white">
+          <Text>Box (backgroundColor=white)</Text>
+        </Box>
+        <Box backgroundColor="white">
+          <Text>Box (backgroundColor=white)</Text>
+        </Box>
+      </Inline>
+    </Box>
   );
+};
+
+Example.args = {
+  spacing: 'md',
 };
