+++ packages/react/src/layout/inline/Inline.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable react/no-array-index-key */
 import * as React from 'react';
-import { get } from 'styled-system';
 
-import { useTheme, Theme } from '../../theme';
-import { Box, BoxProps } from '../box';
+import { Theme } from '../../theme';
+import { BoxProps } from '../box';
+import { Stack } from '../stack';
 
 type Space = keyof Theme['space'];
 
@@ -20,32 +20,14 @@ export interface InlineProps extends Omit<BoxProps, 'color'>, SpacingProps {
 }
 
 /**
- * @deprecated This component will soon be replaced with `<Stack direction="row" />`
+ * @deprecated Please use `<Stack direction="horizontal" />`.
  */
 const Inline = React.forwardRef<HTMLDivElement, InlineProps>(
   ({ children, spacing = 'sm', alignItems, ...rest }, ref) => {
-    const theme = useTheme();
-    const validChildrenArray = React.Children.toArray(children).filter(React.isValidElement);
-
-    const negativeSpacing = React.useMemo(() => `-${get(theme, `space.${spacing}`, 0)}px`, [theme]);
-
     return (
-      <Box ref={ref} marginTop={negativeSpacing} {...rest}>
-        <Box display="flex" flexWrap="wrap" alignItems={alignItems} marginLeft={negativeSpacing}>
-          {validChildrenArray.map((child, i) => {
-            const spacingProps = { mt: spacing, ml: spacing, mb: 0, mr: 0 };
-            if (typeof child === 'string' || child.type === React.Fragment) {
-              return (
-                <Box key={`inline-child-${i}`} {...spacingProps}>
-                  {child}
-                </Box>
-              );
-            }
-
-            return React.cloneElement(child, spacingProps);
-          })}
-        </Box>
-      </Box>
+      <Stack ref={ref} direction="horizontal" spacing={spacing} {...rest}>
+        {children}
+      </Stack>
     );
   }
 );
