+++ packages/react/src/layout/stack/Stack.stories.tsx
@@ -1,65 +1,38 @@
 import * as React from 'react';
+import { Story } from '@storybook/react';
 
-import { SystemBlock, ComponentBlock } from '../../utils/storybook';
-import { Card } from '../../components';
 import { Box } from '../box';
 import { Text } from '../typography';
-import { Stack } from '.';
-
-const readme = require('./README.md');
+import Stack, { StackProps } from './Stack';
 
 export default {
   title: 'Core/Layout/Stack',
   component: Stack,
-  parameters: {
-    notes: { markdown: readme },
-    jsx: { skip: 2 },
+  argTypes: {
+    spacing: {
+      options: ['xxxs', 'xxs', 'xs', 'sm', 'md', 'lg', 'xl', 'xxl'],
+    },
   },
 };
 
-export const Example = () => {
+export const Example: Story<StackProps> = ({ spacing }) => {
   return (
-    <SystemBlock
-      title="Stack"
-      subtitle="Use the Stack component to evenly space inner elements vertically"
-      noBackground
-    >
-      <Stack spacing="xl">
-        <ComponentBlock title="Stack (spacing=md)">
-          <Card padding="sm" elevation={2}>
-            <Box backgroundColor="red02" overflow="hidden">
-              <Stack spacing="md">
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-              </Stack>
-            </Box>
-          </Card>
-        </ComponentBlock>
-        <ComponentBlock title="Stack (spacing=lg)">
-          <Card padding="sm" elevation={2}>
-            <Box backgroundColor="red02" overflow="hidden">
-              <Stack spacing="lg">
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-                <Box backgroundColor="white">
-                  <Text>Box (backgroundColor=white)</Text>
-                </Box>
-              </Stack>
-            </Box>
-          </Card>
-        </ComponentBlock>
+    <Box backgroundColor="red02">
+      <Stack spacing={spacing}>
+        <Box backgroundColor="white">
+          <Text>Box (backgroundColor=white)</Text>
+        </Box>
+        <Box backgroundColor="white">
+          <Text>Box (backgroundColor=white)</Text>
+        </Box>
+        <Box backgroundColor="white">
+          <Text>Box (backgroundColor=white)</Text>
+        </Box>
       </Stack>
-    </SystemBlock>
+    </Box>
   );
+};
+
+Example.args = {
+  spacing: 'md',
 };
