+++ packages/react/src/stories/elevation.stories.mdx
@@ -1,6 +1,7 @@
 import { Meta, Story } from '@storybook/addon-docs';
 import { ElevationGrid, ElevationDetail } from '../utils/storybook';
-import { Box, Card } from '../layout';
+import { Card } from '../components';
+import { Box } from '../layout';
 
 <Meta title="Foundations/Elevation" />
